requirements.txt
absl-py==1.4.0
backcall==0.2.0
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
ConfigArgParse==1.5.5
dataclasses==0.8
decorator==5.1.1
entrypoints==0.4
Flask==2.0.3
future==0.18.3
google-auth==2.21.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
lxml==4.9.3
Markdown==3.3.7
MarkupSafe==2.0.1
nest-asyncio==1.5.6
numpy==1.19.5
oauthlib==3.2.2
OpenNMT-py @ git+https://github.com/stevenkbennett/OpenNMT-py.git@hackathon
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
portalocker==2.7.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.14.0
pyonmttok==1.26.4
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==25.1.0
rdkit-pypi==2021.9.4
regex==2023.6.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.3.1
sentencepiece==0.1.99
six==1.16.0
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.5.1
torchaudio==0.5.1
torchtext==0.5.0
torchvision==0.6.1
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.16
waitress==2.0.0
wcwidth==0.2.6
Werkzeug==2.0.3
zipp==3.6.0
